backend/media-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.7.2
python-multipart==0.0.9
pyjwt==2.9.0
